services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
SQLAlchemy>=2.0.32
psycopg[binary]>=3.2.1
python-dotenv>=1.0.1
celery>=5.4.0
redis>=5.0.0
prometheus-fastapi-instrumentator>=6.1.0
python-jose[cryptography]
passlib[bcrypt]